same thing but with selector class
