+++ .gitignore
@@ -1,7 +1,3 @@
-# Useless files/folders
-my_datapack/
-my_tests.py
-
 # Byte-compiled / optimized / DLL files
 __pycache__/
 *.py[cod]
