+++ __init__.py
@@ -1,3 +1,6 @@
-from .wmcpy.core import *
+from .wmcpy.class.datapack import *
+from .wmcpy.class.file import *
+from .wmcpy.class.workspace import *
+from .wmcpy.class.selector import *
 VERSION = "0.0.1"
 AUTHOR = "Theskyblockman (for this branch)"
