+++ wmcpy/class/workspace.py
@@ -1,0 +1,10 @@
+from .File import File
+
+class Workspace():
+    def __init__(self, name: str) -> None:
+        if name.endswith('.mcfunction'): name.replace('.mcfunction', '', 1)
+        self.name = name
+        self.files = None
+    def add_file(self, file: File):
+        if self.files == None: self.files = []
+        self.files.append(file)
